+++ packages/feeds-client/src/feed/feed.ts
@@ -51,6 +51,7 @@ import type {
   PagerResponseWithLoadingStates,
 } from '../types';
 import { checkHasAnotherPage, Constants, uniqueArrayMerge } from '../utils';
+import { handleActivityMarked } from './event-handlers/activity/handle-activity-marked';
 
 export type FeedState = Omit<
   Partial<GetOrCreateFeedResponse & FeedResponse>,
@@ -180,19 +181,7 @@ export class Feed extends FeedApi {
     'feeds.poll.vote_removed': Feed.noop,
     'feeds.activity.pinned': Feed.noop,
     'feeds.activity.unpinned': Feed.noop,
-    'feeds.activity.marked': (event) => {
-      const currentState = this.currentState;
-      const result = updateNotificationStatusFromActivityMarked(
-        event,
-        currentState.notification_status,
-        currentState.aggregated_activities,
-      );
-      if (result.changed) {
-        this.state.partialNext({
-          ...result.data,
-        });
-      }
-    },
+    'feeds.activity.marked': handleActivityMarked.bind(this),
     'moderation.custom_action': Feed.noop,
     'moderation.flagged': Feed.noop,
     'moderation.mark_reviewed': Feed.noop,
